Refactored to use express router
